+++ Packages/BrandonUtils/CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.19.1] - 9/3/2021
+
+### Changed
+
+- Moved `Mathb` from `BrandonUtils.Runtime` -> `BrandonUtils.Standalone`
+- Removed references to `Vector2` in `Mathb`
+
+### Added
+
+- `Ceil` and `Floor` methods to `Mathb`
+- Fleshed out primitive numeric type overloads in `Mathb` (`Clamp`, `Clamp01`, etc.)
+
 ## [2.18.2] - 9/3/2021
 
 ### Changed